learnflow-app/codereview-service/requirements.txt
# codereview-service/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
groq>=0.11.0
pydantic>=2.8.0
python-dotenv>=1.0.0
